feat: expand CI pipeline (ruff+mypy+bandit+vitest+helm+docker-build+matrix)
Auto-fix 281 ruff issues; 4 remaining F401 noqa-ed. CI now runs:

python (3.10/3.11/3.12): ruff lint + format check + mypy + bandit + pytest

frontend: vitest on web-portal

helm: lint + template render + yaml validation

docker-build: 5 service images + smoke test

all-green gate: aggregate job with fail-fast

+++ services/agent-platform-api/src/ai_employee/agent_platform_api/platform_metrics.py
@@ -98,8 +98,8 @@ def _percentile(values: list[float], p: float) -> float:
         return 0.0
     s = sorted(values)
     rank = (p / 100) * (len(s) - 1)
-    lo = int(math.floor(rank))
-    hi = int(math.ceil(rank))
+    lo = math.floor(rank)
+    hi = math.ceil(rank)
     if lo == hi:
         return float(s[lo])
     return float(s[lo] + (s[hi] - s[lo]) * (rank - lo))
